feat(gui-v2): split into components to make toolbar clean
